+++ Testapp/TestBehaviour.cpp
@@ -11,7 +11,7 @@ void TestBehaviour::OnAwake()
 
 void TestBehaviour::OnUpdate(float _fDeltaTime)
 {
-	m_entity.m_transform.rotation.y = ((sin(glfwGetTime()) * 30.0f));
+	m_entity.m_transform.rotation.y = ((sin((float)glfwGetTime()) * 30.0f));
 	if (m_entity.m_transform.rotation.y > 360.0f)
 	{
 		m_entity.m_transform.rotation.y = 0.0f;
